app/requirements.txt
flask==3.0.3
requests==2.22.0
bs4==0.0.2
lxml==5.3.0
psycopg2-binary==2.9.10
